Implement TCP's timer function
To support TCP's retransmission, timer function
is implemented.

The timer function called 'tcp_slowtimo' is
invoked per 500 ms.(This value could be find in
config.h) Each of the timer counter is tested,
and if non-zero, the counter is decremented.
When the counter reachs 0, 'tcp_timer' is called.

According to the index of counter, 'tcp_timer'
will retransmit segments or  remove a connection
which doesn't enter the 'TCP_ESTABLISHED' state
in time.

To update the value of RTO, 'tcp_rto_update'
is implemented. I refer to 4.4BSD-lite when
writing these code. Chapter 25 of Stevens book
(1) have detailed explanations.

The TCP segment transmitting and acknowledging
routines are still not implemented; I've merely
defined them as placeholders.

(1) "TCP/IP Illustrated, Volume 2: the implementation"
 Gary R. Wright and W. Richard Stevens